backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Nirmal Joji",email = "nirmalsharon.joji@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "langchain-core (>=0.3.41,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "assistant-stream (>=0.0.14,<0.0.15)",
    "motor (>=3.7.0,<4.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langchain (>=0.3.20,<0.4.0)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "chardet (>=5.2.0,<6.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.1,<3.0.0)",
    "langchain-mongodb (>=0.5.0,<0.6.0)",
    "mcp-server-fetch (>=2025.1.17,<2026.0.0)",
    "langfuse (>=2.14.0,<3.0.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "unstructured-ingest[pdf] (>=0.5.13,<0.6.0)",
    "unstructured (>=0.16.25,<0.17.0)",
    "langchain-unstructured (>=0.1.6,<0.2.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.1.1,<3.0.0)",
     "langchain-mcp-adapters (>=0.0.3,<0.0.4)",
     "apscheduler (>=3.11.0,<4.0.0)",
     "notion-client (>=2.3.0,<3.0.0)",
     "langgraph-supervisor (>=0.0.11,<0.0.12)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"